feat: Added typings
