+++ src/domWrapper.ts
@@ -3,6 +3,9 @@ import { isElementVisible } from './utils/isElementVisible'
 import BaseWrapper from './baseWrapper'
 import { createWrapperError } from './errorWrapper'
 import WrapperLike from './interfaces/wrapperLike'
+import { DomEventName } from './constants/dom-event-types'
+import { createDOMEvent, TriggerOptions } from './createDomEvent'
+import { nextTick } from 'vue'
 
 export class DOMWrapper<ElementType extends Element>
   extends BaseWrapper<ElementType>
@@ -136,5 +139,22 @@ export class DOMWrapper<ElementType extends Element>
     }
 
     return new DOMWrapper(parentElement).trigger('change')
+  }
+
+  async trigger(eventString: DomEventName | string, options?: TriggerOptions) {
+    if (options && options['target']) {
+      throw Error(
+        `[vue-test-utils]: you cannot set the target value of an event. See the notes section ` +
+          `of the docs for more details—` +
+          `https://vue-test-utils.vuejs.org/api/wrapper/trigger.html`
+      )
+    }
+
+    if (this.element && !this.isDisabled()) {
+      const event = createDOMEvent(eventString, options)
+      this.element.dispatchEvent(event)
+    }
+
+    return nextTick()
   }
 }
