+++ src/vueWrapper.ts
@@ -17,6 +17,7 @@ import { mergeDeep } from './utils'
 import { emitted, recordEvent } from './emit'
 import BaseWrapper from './baseWrapper'
 import WrapperLike from './interfaces/wrapperLike'
+import { createDOMEvent, TriggerOptions } from './createDomEvent'
 
 export class VueWrapper<T extends ComponentPublicInstance>
   extends BaseWrapper<T['$el']>
@@ -256,6 +257,26 @@ export class VueWrapper<T extends ComponentPublicInstance>
     }
 
     this.__app.unmount()
+  }
+
+  async trigger(
+    eventString: Parameters<T['$emit']>[0],
+    options?: TriggerOptions
+  ) {
+    if (options && options['target']) {
+      throw Error(
+        `[vue-test-utils]: you cannot set the target value of an event. See the notes section ` +
+          `of the docs for more details—` +
+          `https://vue-test-utils.vuejs.org/api/wrapper/trigger.html`
+      )
+    }
+
+    if (this.element && !this.isDisabled()) {
+      const event = createDOMEvent(eventString, options)
+      this.element.dispatchEvent(event)
+    }
+
+    return nextTick()
   }
 }
